some fixes and build in the update function.
currently shown as action menu ... but i fear thats not the best place.
Current update time is again 1 min.
we could use 4 new  settings:
- show_btc (boolean) sets if btc gets updated
- show_ltc (boolean) sets if ltc gets updated
- show_ftc (boolean) sets if ftc gets updated
- update_rate (int) all _milliseconds_ gets updated
did not build them correctly in ... we need to see if we like them on
settings tab or make an settings activity.
i would suggest to use the show_*tc also as indicator if we show them in
the action bar.